# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/protected/message/:messageid": {
            "delete": {
                "description": "delete message with id messageid if you are admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth message"
                ],
                "summary": "delete messaage by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/protected/message/:serverid": {
            "post": {
                "description": "send a message to server with serverid if you are a member",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth message"
                ],
                "summary": "send message to a server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "message",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/protected/message/byserver/:serverid": {
            "get": {
                "description": "get message by serverid if you are sever member or admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth message"
                ],
                "summary": "get message by server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/protected/message/byuser/:userid": {
            "get": {
                "description": "get message by userid if you are same user or admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth message"
                ],
                "summary": "get message by user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/protected/server/:serverid": {
            "delete": {
                "description": "delete a server if you are admin or server admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth server"
                ],
                "summary": "delete a server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/protected/server/:serverid/:userid": {
            "delete": {
                "description": "kick user with userid from server if you are server admin, admin or user in question",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth server"
                ],
                "summary": "kick a user from server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/protected/server/:serverid/adduser": {
            "post": {
                "description": "add user to server if you are server admin or admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth server"
                ],
                "summary": "add user to server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "user id to add",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/protected/server/:serverid/name": {
            "patch": {
                "description": "change server name if you are admin or server admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth server"
                ],
                "summary": "change server name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "new server name",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/protected/server/:serverid/users": {
            "get": {
                "description": "get users on a server if you are a member of that server or admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth server"
                ],
                "summary": "get users on a server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/protected/server/new": {
            "post": {
                "description": "create a new server and add creating user as server admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth server"
                ],
                "summary": "create a new server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "server name",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}